--- a/.dockerignore
+++ b/.dockerignore
@@ -1,0 +1,3 @@
+db/*
+vendor/*
+.git