--- a/.gitignore
+++ b/.gitignore
@@ -1,2 +1,3 @@
 db/*.sqlite3*
+db/trends
 .bundle