--- a/.gitignore
+++ b/.gitignore
@@ -1,3 +1,4 @@
 db/*.sqlite3*
 db/trends
+db/test_trends
 .bundle