--- a/Gemfile
+++ b/Gemfile
@@ -13,3 +13,7 @@ gem "minitest"
 gem "rufo"
 gem "leveldb-ruby"
 gem "msgpack"
+
+group :development do
+  gem "pry"
+end