--- a/bin/console
+++ b/bin/console
@@ -1,0 +1,6 @@
+#!/usr/bin/env ruby
+
+require_relative '../lib/database'
+require 'pry'
+
+binding.pry