--- a/config/database.yml.erb
+++ b/config/database.yml.erb
@@ -3,7 +3,7 @@ test:
     adapter: "sqlite"
     timeout: 60000
   leveldb:
-    dir: "db/trends"
+    dir: "db/test_trends"
 
 development:
   db: