--- a/docker/bestgems-app/Dockerfile
+++ b/docker/bestgems-app/Dockerfile
@@ -5,6 +5,8 @@ ADD docker/bin/wait-for-it.sh /bin
 
 WORKDIR /srv/bestgems.org
 
+RUN apt-get update && apt-get -y install libleveldb-dev && apt-get clean
+
 RUN bundle install --path .bundle
 
 CMD bundle exec rackup -o 0.0.0.0