--- a/lib/batch/dummy.rb
+++ b/lib/batch/dummy.rb
@@ -1,0 +1,3 @@
+require_relative '../database'
+
+puts Time.now.to_s