--- a/lib/models/trend.rb
+++ b/lib/models/trend.rb
@@ -1,86 +1,96 @@
-class Trend
-  LEVELDB_DIR = Settings.leveldb['dir']
+TREND_DRB_URI = 'druby://localhost:16330'
 
-  @@level_db = LevelDB::DB.new LEVELDB_DIR
+if ENV['BESTGEMS_TREND_SERVER'] == 'true'
+  class Trend
+    include DRb::DRbUndumped
 
-  def self.get_a(key)
-    unpack(@@level_db[key])
-  end
+    LEVELDB_DIR = Settings.leveldb['dir']
 
-  def self.put_a(key, td_list)
-    @@level_db[key] = pack(td_list) if td_list
-  end
+    @@level_db = LevelDB::DB.new LEVELDB_DIR
 
-  def self.delete_a(key)
-    @@level_db.delete(key)
-  end
+    def self.get_a(key)
+      unpack(@@level_db[key])
+    end
 
-  def self.all(&blk)
-    @@level_db.each{|k, v|
-      blk.call(k, unpack(v))
-    }
-  end
+    def self.put_a(key, td_list)
+      @@level_db[key] = pack(td_list) if td_list
+    end
 
-  def self.get(gem_id)
-    from = "#{gem_id}.0"
-    to = "#{gem_id}.9"
+    def self.delete_a(key)
+      @@level_db.delete(key)
+    end
 
-    td_list = []
+    def self.all(&blk)
+      @@level_db.each{|k, v|
+        blk.call(k, unpack(v))
+      }
+    end
 
-    @@level_db.each(from: from, to: to){|_, value|
-      td_list += unpack(value) if value
-    }
-    
-    td_list.sort_by{|td| td.date }
-  end
+    def self.get(gem_id)
+      from = "#{gem_id}.0"
+      to = "#{gem_id}.9"
 
-  def self.put(gem_id, *td_list)
-    raise ArgumentError if td_list.size == 0
+      td_list = []
 
-    td_list.group_by{|td| td.key(gem_id) }.each{|key, monthly_td_list|
-      update(key, monthly_td_list)
-    }
-  end
+      @@level_db.each(from: from, to: to){|_, value|
+        td_list += unpack(value) if value
+      }
+      
+      td_list.sort_by{|td| td.date }
+    end
 
-  def self.update(key, td_list)
-    if exist_td_list = get_a(key)
-      merged_td_list = merge_td_list(exist_td_list, td_list)
+    def self.put(gem_id, *td_list)
+      raise ArgumentError if td_list.size == 0
 
-      put_a(key, merged_td_list)
-    else
-      put_a(key, td_list)
+      td_list.group_by{|td| td.key(gem_id) }.each{|key, monthly_td_list|
+        update(key, monthly_td_list)
+      }
+    end
+
+    def self.update(key, td_list)
+      if exist_td_list = get_a(key)
+        merged_td_list = merge_td_list(exist_td_list, td_list)
+
+        put_a(key, merged_td_list)
+      else
+        put_a(key, td_list)
+      end
+    end
+
+    def self.merge_td_list(exist_td_list, new_td_list)
+      hash = {}
+
+      exist_td_list.each{|td|
+        hash[td.date] = td
+      }
+
+      new_td_list.each{|td|
+        hash[td.date] = td
+      }
+
+      hash.values
+    end
+
+    def self.empty?
+      @@level_db.each{|_|
+        return false
+      }
+
+      true
+    end
+
+    private
+
+    def self.unpack(value)
+      MessagePack.unpack(value) if value
+    end
+
+    def self.pack(td_list)
+      MessagePack.pack(td_list)
     end
   end
 
-  def self.merge_td_list(exist_td_list, new_td_list)
-    hash = {}
-
-    exist_td_list.each{|td|
-      hash[td.date] = td
-    }
-
-    new_td_list.each{|td|
-      hash[td.date] = td
-    }
-
-    hash.values
-  end
-
-  def self.empty?
-    @@level_db.each{|_|
-      return false
-    }
-
-    true
-  end
-
-  private
-
-  def self.unpack(value)
-    MessagePack.unpack(value) if value
-  end
-
-  def self.pack(td_list)
-    MessagePack.pack(td_list)
-  end
+  DRb.start_service(TREND_DRB_URI, Trend, safe_level: 1)
+else
+  Trend = DRbObject.new_with_uri(TREND_DRB_URI)
 end