--- a/lib/models/trend_data.rb
+++ b/lib/models/trend_data.rb
@@ -1,21 +1,21 @@
 class TrendData
-  attr_accessor :date, :total_downloads, :total_rank, :daily_downloads, :daily_rank
+  attr_accessor :date, :total_downloads, :total_ranking, :daily_downloads, :daily_ranking
 
   def self.msgpack_type
     0x01
   end
 
   def to_msgpack_ext
-    MessagePack.pack [@date.jd, @total_downloads, @total_rank, @daily_downloads, @daily_rank]
+    MessagePack.pack [@date.jd, @total_downloads, @total_ranking, @daily_downloads, @daily_ranking]
   end
 
   def self.from_msgpack_ext(data)
-    jd, total_downloads, total_rank, daily_downloads, daily_rank = MessagePack.unpack(data)
-    self.new(Date.jd(jd), total_downloads, total_rank, daily_downloads, daily_rank)
+    jd, total_downloads, total_ranking, daily_downloads, daily_ranking = MessagePack.unpack(data)
+    self.new(Date.jd(jd), total_downloads, total_ranking, daily_downloads, daily_ranking)
   end
 
-  def initialize(date, total_downloads, total_rank, daily_downloads, daily_rank)
-    @date, @total_downloads, @total_rank, @daily_downloads, @daily_rank = date, total_downloads, total_rank, daily_downloads, daily_rank
+  def initialize(date, total_downloads, total_ranking, daily_downloads, daily_ranking)
+    @date, @total_downloads, @total_ranking, @daily_downloads, @daily_ranking = date, total_downloads, total_ranking, daily_downloads, daily_ranking
   end
 
   def key(gem_id)
@@ -25,9 +25,9 @@ class TrendData
   def ==(other)
     @date == other.date \
       && @total_downloads == other.total_downloads \
-      && @total_rank == other.total_rank \
+      && @total_ranking == other.total_ranking \
       && @daily_downloads == other.daily_downloads \
-      && @daily_rank == other.daily_rank
+      && @daily_ranking == other.daily_ranking
   end
 
   MessagePack::DefaultFactory.register_type(self.msgpack_type, self, packer: :to_msgpack_ext, unpacker: :from_msgpack_ext)