--- a/lib/models/trend_data_set.rb
+++ b/lib/models/trend_data_set.rb
@@ -1,0 +1,17 @@
+class TrendDataSet
+  def initialize(td_list)
+    @sorted_td_list = td_list.sort_by{|td| td.date }
+  end
+
+  def downloads_trends
+    @sorted_td_list.map{|td|
+      {date: td.date, total_downloads: td.total_downloads, daily_downloads: td.daily_downloads}
+    }
+  end
+
+  def ranking_trends
+    @sorted_td_list.map{|td|
+      {date: td.date, total_ranking: td.total_ranking, daily_ranking: td.daily_ranking}
+    }
+  end
+end