--- a/tools/archive.rb
+++ b/tools/archive.rb
@@ -100,7 +100,9 @@ SQL
   end
 end
 
-min_date = Date.parse([Value.min(:date), Ranking.min(:date)].min)
+
+min_data = [Value.min(:date), Ranking.min(:date)].min
+min_date = Date.parse(min_data) unless min_data.is_a? Date
 min_date = min_date - (min_date.day - 1)
 
 last_archive_date = Date.today - 32