--- a/tools/archive.rb
+++ b/tools/archive.rb
@@ -101,7 +101,7 @@ SQL
 end
 
 
-min_data = [Value.min(:date), Ranking.min(:date)].min
+min_date = [Value.min(:date), Ranking.min(:date)].min
 min_date = Date.parse(min_data) unless min_data.is_a? Date
 min_date = min_date - (min_date.day - 1)