--- a/tools/archive.rb
+++ b/tools/archive.rb
@@ -102,7 +102,7 @@ end
 
 
 min_date = [Value.min(:date), Ranking.min(:date)].min
-min_date = Date.parse(min_data) unless min_data.is_a? Date
+min_date = Date.parse(min_date) unless min_date.is_a? Date
 min_date = min_date - (min_date.day - 1)
 
 last_archive_date = Date.today - 32