… .travis.yml
… Gemfile.lock
… docker/docker-compose.yml
… test/batch/test_update_daily_downloads.rb
… test/batch/test_update_daily_ranking.rb
… test/batch/test_update_featured_ranking.rb
… test/batch/test_update_featured_score.rb
… test/batch/test_update_gems.rb
… test/batch/test_update_statistics_daily_downloads.rb
… test/batch/test_update_statistics_num_of_gems.rb
… test/batch/test_update_statistics_total_downloads.rb
… test/batch/test_update_total_downloads.rb
… test/batch/test_update_total_ranking.rb
… test/batch/test_update_trends.rb
… test/models/test_gems.rb
… test/models/test_model.rb
… test/models/test_ranking.rb
… test/models/test_trend.rb
… test/models/test_trend_data.rb
… test/models/test_trend_data_set.rb
… test/test_access_tables.rb
… test/test_helper.rb
… tools/export_scraped_data.rb
… tools/generate_dummy_data.rb
… tools/import_scraped_data.rb
… views/search.erb