# Summary

Ref: https://github.com/xmisao/bestgems.org/issues/32

## Effect

* 3 times faster
* 3 times concurrency
* Stable performance
* Database capacity reduction

## Means

* Store time series data to LevelDB
* Move old trend data to archiving tables

# Siege benchmarks

On AWS t2.micro instance. (1vCPU, 1GB RAM)

## Before

```
$ siege -b -c 1 -r 10 -f before_url.txt 
[alert] Zip encoding disabled; siege requires zlib support to enable it
** SIEGE 4.0.2
** Preparing 1 concurrent users for battle.
The server is now under siege...
Transactions:                     40 hits
Availability:                 100.00 %
Elapsed time:                  12.16 secs
Data transferred:               1.10 MB
Response time:                  0.30 secs
Transaction rate:               3.29 trans/sec
Throughput:                     0.09 MB/sec
Concurrency:                    1.00
Successful transactions:          20
Failed transactions:               0
Longest transaction:            1.53
Shortest transaction:           0.01
 
$ siege -b -c 2 -r 10 -f before_url.txt 
[alert] Zip encoding disabled; siege requires zlib support to enable it
** SIEGE 4.0.2
** Preparing 2 concurrent users for battle.
The server is now under siege...
Transactions:                     80 hits
Availability:                 100.00 %
Elapsed time:                  25.47 secs
Data transferred:               2.21 MB
Response time:                  0.64 secs
Transaction rate:               3.14 trans/sec
Throughput:                     0.09 MB/sec
Concurrency:                    1.99
Successful transactions:          40
Failed transactions:               0
Longest transaction:            4.20
Shortest transaction:           0.01

$ siege -b -c 4 -r 10 -f before_url.txt 
[alert] Zip encoding disabled; siege requires zlib support to enable it
** SIEGE 4.0.2
** Preparing 4 concurrent users for battle.
The server is now under siege...
Transactions:                    160 hits
Availability:                 100.00 %
Elapsed time:                  38.79 secs
Data transferred:               4.36 MB
Response time:                  0.97 secs
Transaction rate:               4.12 trans/sec
Throughput:                     0.11 MB/sec
Concurrency:                    3.99
Successful transactions:          80
Failed transactions:               0
Longest transaction:            4.18
Shortest transaction:           0.01

$ siege -b -c 8 -r 10 -f before_url.txt 
[alert] Zip encoding disabled; siege requires zlib support to enable it
** SIEGE 4.0.2
** Preparing 8 concurrent users for battle.
The server is now under siege...
Transactions:                    320 hits
Availability:                 100.00 %
Elapsed time:                 100.00 secs
Data transferred:               8.83 MB
Response time:                  2.46 secs
Transaction rate:               3.20 trans/sec
Throughput:                     0.09 MB/sec
Concurrency:                    7.88
Successful transactions:         160
Failed transactions:               0
Longest transaction:           11.93
Shortest transaction:           0.01
```

## After

```
$ siege -b -c 1 -r 10 -f after_url.txt 
[alert] Zip encoding disabled; siege requires zlib support to enable it
** SIEGE 4.0.2
** Preparing 1 concurrent users for battle.
The server is now under siege...
Transactions:                     40 hits
Availability:                 100.00 %
Elapsed time:                   4.97 secs
Data transferred:               1.10 MB
Response time:                  0.12 secs
Transaction rate:               8.05 trans/sec
Throughput:                     0.22 MB/sec
Concurrency:                    1.00
Successful transactions:          20
Failed transactions:               0
Longest transaction:            0.37
Shortest transaction:           0.01

$ siege -b -c 2 -r 10 -f after_url.txt 
[alert] Zip encoding disabled; siege requires zlib support to enable it
** SIEGE 4.0.2
** Preparing 2 concurrent users for battle.
The server is now under siege...
Transactions:                     80 hits
Availability:                 100.00 %
Elapsed time:                   7.60 secs
Data transferred:               2.19 MB
Response time:                  0.19 secs
Transaction rate:              10.53 trans/sec
Throughput:                     0.29 MB/sec
Concurrency:                    1.98
Successful transactions:          40
Failed transactions:               0
Longest transaction:            0.68
Shortest transaction:           0.01

$ siege -b -c 4 -r 10 -f after_url.txt 
[alert] Zip encoding disabled; siege requires zlib support to enable it
** SIEGE 4.0.2
** Preparing 4 concurrent users for battle.
The server is now under siege...
Transactions:                    160 hits
Availability:                 100.00 %
Elapsed time:                  13.58 secs
Data transferred:               4.33 MB
Response time:                  0.33 secs
Transaction rate:              11.78 trans/sec
Throughput:                     0.32 MB/sec
Concurrency:                    3.89
Successful transactions:          80
Failed transactions:               0
Longest transaction:            1.31
Shortest transaction:           0.01

$ siege -b -c 8 -r 10 -f after_url.txt 
[alert] Zip encoding disabled; siege requires zlib support to enable it
** SIEGE 4.0.2
** Preparing 8 concurrent users for battle.
The server is now under siege...
Transactions:                    320 hits
Availability:                 100.00 %
Elapsed time:                  25.83 secs
Data transferred:               8.77 MB
Response time:                  0.63 secs
Transaction rate:              12.39 trans/sec
Throughput:                     0.34 MB/sec
Concurrency:                    7.85
Successful transactions:         160
Failed transactions:               0
Longest transaction:            2.59
Shortest transaction:           0.01
```